kusin taas jotai -> korjasin

+++ server/server_functions.c
@@ -244,7 +244,7 @@ int gameMsgPacker(char **pPL, Game *pGame, int toPlayerID, int msgSubType,
 			*(uint16_t *) &pPL[ind] = htons(pPlayer->direction[1]);
 			ind += sizeof(uint16_t);
 			*(uint32_t *) &pPL[ind] = htonl(pPlayer->scale);
-			ind += sizeof(uint16_t);
+			ind += sizeof(uint32_t);
 
 			indNPla = ind;  // store index of nPlayers
 			ind += sizeof(uint8_t);
